+++ src/helpers/setMaxZoomLevel.js
